Debug Java Script - Corrected - LEWIS
I found what I thought was one error. On line 36 and 37, I saw: 
36: const rate;
37: const rate = document.getElementById("rate").value / 100;

Line 36 didn't look right so I deleted it and made a note of it within the java file.

I ran the script again through Chrome and was prompted with a new error. Invalid token.  Off to the side of the error in the console, it said there was an error on line 39. When I checked there, I found the "years" error. It was missing a ".
After fixing the quotation, I ran the script again and found no more errors.
I made notes within the file of all my changes.